catalog/stacks/ms-office/python/requirements.txt
mcp>=1.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
